+++ mRemoteNG/Properties/AssemblyInfo.cs
@@ -11,7 +11,7 @@ using System.Resources;
 
 // Compute version values
 
-//Build nr: 3225
+//Build nr: 3388
 
 // General Information
 [assembly: AssemblyTitle("mRemoteNG")]
@@ -19,12 +19,12 @@ using System.Resources;
 [assembly: AssemblyConfiguration("x64")]
 [assembly: AssemblyCompany("Profi-KOM Ltd.")]
 [assembly: AssemblyProduct("mRemoteNG Connection Manager")]
-[assembly: AssemblyCopyright("(c) 2025 mRemoteNG")]
+[assembly: AssemblyCopyright("(c) 2026 mRemoteNG")]
 [assembly: AssemblyTrademark("Profi-KOM LTd.")]
 [assembly: AssemblyCulture("")]
 
 // Version information
-[assembly: AssemblyVersion("1.78.2.3225")]
-[assembly: AssemblyFileVersion("1.78.2.3225")]
+[assembly: AssemblyVersion("1.80.0.3388")]
+[assembly: AssemblyFileVersion("1.80.0.3388")]
 [assembly: NeutralResourcesLanguageAttribute("en-US")]
-[assembly: AssemblyInformationalVersion("1.78.2 (Nightly Build 3225) x64")]
+[assembly: AssemblyInformationalVersion("1.80.0 (Release 3388) x64")]